Finished save_and_load.py and updated .gitignore.
Added main function to save_and_load.py so it works at all (argparse
doesn't call functions by itself), tested it and fixed an issue where
relative paths would be written to the data_dir and an issue with the
initial check for paths that contained a misplaced variable. Also
updated .gitignore to include testing/ and .vscode/ (I needed to edit
launch.json to debug stuff).